secondsToRecordIR is now a float value. It is initialized with the duration of the room's BRIR

+++ src/SoundSource.cpp
@@ -95,11 +95,11 @@ void SoundSource::setInitialized()
 }
 
 
-void SoundSource::startRecordOfflineOfImpulseResponse(int secondsToRecord) {
+void SoundSource::startRecordOfflineOfImpulseResponse(float secondsToRecord) {
 	
 	samplesVectorCopy = samplesVector;			// Save initial wav file
 	//samplesVector.resize(44100*secondsToRecord);    
-	samplesVector.resize( sampleRate * secondsToRecord);
+	samplesVector.resize( int(sampleRate * secondsToRecord));
 
 	std::fill(samplesVector.begin(), samplesVector.end(), 0.0);   
 	samplesVector.at(1) = 0.99;
